implemented not logged in view of navbar
