+++ frontend/src/app/ContactUs/page.tsx
@@ -1,5 +1,5 @@
 import Contact from "../../components/Contact";
-import Navbar from "../../components/Navbar";
+import Navbar from "../../components/VolunteerNavbar";
 
 export default function ContactUsPage() {
   return (
