+++ frontend/src/components/Navbar.tsx
@@ -1,46 +1,46 @@
 "use client";
 
 import Link from "next/link";
-import { usePathname } from "next/navigation";
 import { useState } from "react";
 import styles from "../styles/navbar.module.css";
 
 export default function Navbar() {
-  const pathname = usePathname();
-  const [collapsed, setCollapsed] = useState(false);
-
-  const navItems = [
-    { label: "Home", href: "/" },
-    { label: "My Profile", href: "/Profile" },
-    { label: "Volunteer", href: "/Volunteer" },
-    // don't have events implemented
-    { label: "My Events", href: "/MyEvents" },
-    { label: "Upcoming Events", href: "/Events/Upcoming" },
-    { label: "Past Events", href: "/Events/Past" },
-    { label: "About Us", href: "/About" },
-    { label: "Contact", href: "/ContactUs" },
-    // donate also not implemented
-    { label: "Donate", href: "/Donate" },
-  ];
+  const [open, setOpen] = useState(false);
 
   return (
-    <nav className={`${styles.sidebar} ${collapsed ? styles.collapsed : ""}`}>
-      <button className={styles.toggleButton} onClick={() => setCollapsed(!collapsed)}>
-        ☰
-      </button>
-      <ul className={styles.navList}>
-        {navItems.map((item) => (
-          <li key={item.href}>
-            <Link href={item.href} className={`${styles.navLink} ${pathname === item.href ? styles.active : ""}`}>
-              <span className={styles.linkText}>{item.label}</span>
-            </Link>
-          </li>
-        ))}
-      </ul>
-      <div className={styles.signOut}>
-        <button className={styles.signOutButton}>
-          <span className={styles.linkText}>Sign Out</span>
+    <nav className={styles.navbar}>
+      <Link href="/" className={styles.link}>
+        Home
+      </Link>
+
+      <Link href="/About" className={styles.link}>
+        About Us
+      </Link>
+
+      <Link href="/" className={styles.logo}>
+        Logo
+      </Link>
+
+      <Link href="/ContactUs" className={styles.link}>
+        Contact Us
+      </Link>
+
+      <div className={styles.dropdownWrapper}>
+        <button className={styles.moreButton} onClick={() => setOpen(!open)}>
+          More
         </button>
+
+        <div className={`${styles.dropdown} ${open ? styles.show : ""}`}>
+          <Link href="/Auth/Login" className={styles.dropdownItem}>
+            Sign Up / Login
+          </Link>
+          <Link href="/Events" className={styles.dropdownItem}>
+            Events
+          </Link>
+          <Link href="/Donate" className={styles.dropdownItem}>
+            Donate
+          </Link>
+        </div>
       </div>
     </nav>
   );
