+++ frontend/src/styles/volunteernavbar.module.css
@@ -1,0 +1,122 @@
+.sidebar {
+  width: 200px;
+  min-height: 100vh;
+  box-sizing: border-box;
+  background-color: #ffffff;
+  border-right: 1px solid #e0e0e0;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  padding: 2rem 0;
+  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+  overflow: hidden;
+  transition: width 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.navList {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.navLink {
+  display: block;
+  padding: 0.75rem 1.5rem;
+  color: #333;
+  text-decoration: none;
+  font-size: 1rem;
+  transition: background-color 0.2s ease;
+}
+
+.navLink:hover {
+  background-color: #f5f5f5;
+}
+
+.active {
+  background-color: #e6e6e6;
+  font-weight: 600;
+  border-left: 3px solid #333;
+}
+
+.signOut {
+  padding: 1rem 1.5rem;
+  border-top: 1px solid #e0e0e0;
+}
+
+.signOutButton {
+  width: 100%;
+  padding: 0.5rem 1rem;
+  background-color: transparent;
+  border: none;
+  color: #333;
+  font-size: 1rem;
+  text-align: left;
+  cursor: pointer;
+  font-family: inherit;
+}
+
+.signOutButton:hover {
+  background-color: #f5f5f5;
+}
+
+.collapsed {
+  width: 60px;
+  border-right: none;
+  transition-delay: 0.1s;
+}
+
+.navLink {
+  display: block;
+  padding: 0.75rem 1.5rem;
+  color: #333;
+  text-decoration: none;
+  font-size: 1rem;
+  transition:
+    background-color 0.2s ease,
+    opacity 0.2s ease;
+}
+
+.collapsed .navLink {
+  opacity: 0;
+}
+
+.toggleButton {
+  width: 35px; /* square width */
+  height: 35px; /* square height */
+  display: flex; /* center the ☰ */
+  align-items: left;
+  justify-content: center;
+  font-size: 1.5rem;
+  color: #666666;
+  background-color: #f5f5f5; /* optional background */
+  border: 1px solid #ddd; /* optional border */
+  border-radius: 10px; /* optional rounded corners */
+  cursor: pointer;
+  margin: 0 0 0.5rem 0.5rem; /* small margin from top-left */
+  padding: 0;
+}
+
+.toggleButton:hover {
+  background-color: #e0e0e0;
+}
+
+.linkText {
+  display: inline-block;
+  white-space: nowrap;
+  opacity: 0;
+  transition: opacity 0.2s ease 0.3s;
+}
+
+.sidebar:not(.collapsed) .linkText {
+  opacity: 1;
+  position: static; /* back into normal flow */
+  transition: opacity 0.2s ease 0.3s; /* delay to appear after sidebar expands */
+}
+
+.collapsed .linkText {
+  opacity: 0;
+  pointer-events: none;
+  transition:
+    opacity 0.2s ease 0s,
+    left 0s ease 0s;
+}
